tests/wpt/web-platform-tests/html/semantics/forms/textfieldselection/selection-after-content-change.html

<!doctype html>
<meta charset="utf-8">
<title>Selection indices after content change</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(function() {
    var input = document.createElement("input");
    input.focus();
    input.value = "something something something dark side";
    input.setSelectionRange(4,20);
    input.value = "";
    assert_true(input.selectionStart == input.selectionEnd);
}, "Selection indices after reseting content");
</script>
